third_party/WebKit/LayoutTests/external/wpt/workers/name-property-expected.txt

This is a testharness.js-based test.
FAIL name property value for DedicatedWorkerGlobalScope assert_true: property exists on the global expected true got false
FAIL name property value for SharedWorkerGlobalScope assert_equals: expected "my name" but got "[object Object]"
Harness: the test ran to completion.
